requirements.txt
flask
psycopg2-binary
sqlalchemy